+++ CHANGELOG.md
@@ -1,11 +1,20 @@
 # Changelog
 
-## 0.2.1
+## 0.3.0
 
 ### Build
+* Update go-rockset-client SDK
+* Update grafana plugin SDK
+
+### Code
+* Adapt code to use recent version of the grafana plugin SDK
+
+## 0.2.1
+
+### Code
 * Fixed issues with ordering of values, we now respect the order specified in the query
 
-## 0.2.0 
+## 0.2.0
 
 ### Build
 * Update go-rockset-client SDK